move pd df printing to utils
